Add Widevine CDM component installer.
This CL adds a component installer for Widevine CDM on Windows and Mac. Note that it uses a fake extension ID which will be replaced when we have the real one. Also the component is not registered in chrome_browser_main.cc yet. We need to have the real CDMs uploaded to the server to fully test and enable this feature.

BUG=180260
TEST=Tested local operations (registering plugin, remove old versions, etc) on Linux without sandbox. Will need to test against the server later.

Review URL: https://chromiumcodereview.appspot.com/14192012