Use more than one background windows for a panel stack
The underlying background window is also used to provide the shadow while the foreground panel window does not show the shadow. This allows us to provide the consistent shadow experience no matter which panel in the stack is on focus. Thus the underlying background window is not transparent any more.

Since we might not have fully aligned stacked panels at certain moment, i.e. some panels are being dragged, we use a secondary background window when this is the case.
1) NativePanelStackWindow interface is updated with new methods to support interaction between multiple stack windows. It also adds the support for batch bounds update that allows the smooth animation and transition.
2) StackedPanelCollection now has 2 background windows. The secondary background window is created and destructed on demand.

BUG=223155,177770,177355
TEST=existing tests

Review URL: https://chromiumcodereview.appspot.com/14493019