Turn off Synced Notifications by default
This will provide us more control over when we turn the feature on for which
branch going forwards.

Also provide a three state flag in chrome://flags to enable and disable synced
notifications, and command line flags to override the default for both enabling
and disabling.

BUG=236045
TBR=dimich@chromium.org

Review URL: https://codereview.chromium.org/14520019