Refactor RulesRegistryStorageDelegate, (aggregate into rules registry + separate objects by threads)
We are splitting RulesRegistryStorageDelegate into two parts, one (the outer) working on the rules registry thread, the other (inner) on the UI thread. This separation has been there already into some extent, but we make it complete, and get rid of holding the inner part in a scoped_refptr.

Also, the storage delegate now becomes a part of the RulesRegistryWithCache, as it was the only implementation of
RulesRegistryWithCache::Delegate, and the aggregation trims down the code size and increases clarity.
BUG=218451
R=bauerb@chromium.org, jochen@chromium.org, jyasskin@chromium.org

Review URL: https://codereview.chromium.org/13646014