'Undo Delete' is not functioning as intended under Recent folder of Bookmark Manager.
The Recent list allows the bookmark order to differ from the order
that the bookmarks appear in the bookmark bar or folder. This
is the reason the problem is occuring within the Recent list and not
elsewhere.

If multiple bookmarks are deleted it is required that bookmarks are
restored in the order that they appear. Below is an explaination on
why this is so.

In the scenario detailed in the bug the bookmarks deleted have
|index| values of 0 and 1 (first and second bookmarks in the bookmark
bar). However the bookmark with the index of 1 is restored first,
because first bookmark in the Recent list even though it is the
second bookmark in the bookmark bar. When the deletion operation is
undone the bookmark with |index| 1 is restored first, which results
to a call to create a bookmark at position 1 but the bookmark bar is
empty so that request results in an index out of bounds error.

Note, if there were more bookmark left after the deletion then undo
would have restored both bookmarks however those bookmarks would not
be in their proper position (they would be in position 0 and 2 not
0 and 1).

The solution is to delete the bookmarks in the order of the |index|
variable which will result in the undo to restore those bookmarks in
the correct order.

TEST=Follow the steps on the bug ensuring the order the bookmarks are listed is not the same as they appear on the bookmark bar.
BUG=173981
R=arv@chromium.org

Review URL: https://chromiumcodereview.appspot.com/13868009