Deflake SimpleCache::Close()
It turns out that Close() takes a small, finite amount of time before
it is safe to open an entry again, because it must write the EOF
record before closing the file.

To reduce flakiness, do not permit OpenEntry to proceed while a Close
is pending. This is done by placing a more detailed state machine in the
SimpleEntryImpl, along with a map of all active entries in the SimpleBackendImpl.

Not all flakes are gone (especially around mass dooms), but this should be a strict
improvement.

R=pliard,pasko,felipeg
BUG=233871,236322

Review URL: https://chromiumcodereview.appspot.com/14354008