Android: do not allow vsync subscriptions after ContentViewCore has been destroyed
Do not allow vsync (un)subscriptions ContentViewCore.destroy() has been
called. This is to make sure we keep a balanced count of vsync
subscribers even if a subscription message arrives after the view has
been destroyed.

This patch also refactors the VSyncManager.Provider interface to be
registration-oriented to make it more obvious that registrations and
unregistrations to the vsync signal need to be balanced.

BUG=235623

Review URL: https://chromiumcodereview.appspot.com/13877033