On Windows 8 don't assume that the browser is shutting down in desktop mode because the ASH environment could be still active.
We currently have a check for this in the ShouldStartShutdown function in the application_lifetime.cc file which
is called when a browser window is being closed. It checks for whether both browser lists (desktop and ASH) are
empty. However ASH could be active without any active browser windows.

Updated the checks to account for this.

BUG=235654
R=cpu
TBR=ben

Review URL: https://codereview.chromium.org/14273040