Fix potential deadlock in gpu process creation
When the GPU process crashes, the renderer compositor will try to
recreate the context. In this case, it is possible for it to try to
connect to an already invalid GpuProcessHost, where there are some
race conditions.

Make sure IPC replies to GpuProcessHost::EstablishGpuChannel and
GpuProcessHost::CreateViewCommandBuffer are always called.

BUG=235942

Review URL: https://chromiumcodereview.appspot.com/14522025