[FileAPI] Disable usage cache for origin after file opening for pepper.
For PERSISTENT FileSystem on UnlimitedStorage origin, a file descriptor opened by CreateOrOpen can be passed to a pepper plugin, and we'll lose report-based usage tracking.
This CL conservatively disables usage cache for the origin until browser restart in the case, and invokes the usage recalculation on a usage query.

BUG=220029

Review URL: https://chromiumcodereview.appspot.com/14612002