Add ScopedExtensionElevation to make uninstall of apps work for a managed user.
When clicking with the right mouse button on an app on the
new tab page, the context menu item "Remove from Chrome" is
currently disabled for a managed user. By adding temporary
elevation tied to the scope where UserMayModifySettings is
called, the menu entry will be enabled for a managed user,
and when uninstalling, the extension uninstall confirmation
dialog makes sure to show the passphrase dialog where the
user has to authenticate as the custodian to make the
uninstall successful.

BUG=236367

Review URL: https://chromiumcodereview.appspot.com/14540002