Make UserCloudPolicyManagerChromeOS a ProfileKeyedService.
This enables multi-profile user policy on ChromeOS, and removes the policy code
out of login_utils and chrome_browser_main_chromeos. The blocking initialization
of managed Profiles is now handled by the UserCloudPolicyManagerChromeOS itself,
which is a PKS.

The UserCloudPolicyTokenForwarder PKS exists so that the UCPM doesn't have to
depend on the TokenService, to avoid a circular dependency.

This is a work in progress; the next step is to move the policy code out of
Profile and ProfileImpl into a ProfilePolicyConnector. The logic in
BPC::CreatePolicyServiceForProfile will also go into that new PKS.

BUG=230349, 143187

Review URL: https://chromiumcodereview.appspot.com/14050015