Add DefaultNetworkChanged signal for ConnectivityStateHelper observers
NetworkLibrary's OnNetworkManagerChanged signal is generated even for
changes on the default network (without a corresponding change on
the NetworkManager). Many consumers (e.g. NetworkScreen) rely on this
behavior to not miss connectivity state updates.

Add an explicit DefaultNetworkChanged signal for ConnectivityStateHelper
observers to get signaled for such events when using the NetworkStateHandler
based implementation (which is the default now).

BUG=236417
TEST=1) Start in OOBE state (no Shill default profile, no Chrome local state).
     2) Observer "Continue" button is grayed out.
     3) Choose a wireless network (say "GoogleGuest")
     4) "Continue" button should be come enabled now.

Review URL: https://chromiumcodereview.appspot.com/14587002