Linux sandbox: more paranoid checks at initialization
We're worried that InitializeSandbox() will be called multithreaded, so we
make sure to keep a file descriptor to /proc around, in Debug mode, to be
able to count threads.

Now that seccomp-legacy has been removed, we also simplify the code around
pre-initialization.

BUG=162334
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/14411008