cc: Remove unused is_live flag from TilePriority
After danakj's relanding of whunt's change to PictureLayerTiling, any
tile that would have been marked is_live=false is now just not created
at all.  The upside is also that the live_or_allocated_tiles_ temporary
vector goes away.  Because we need to be able to sort all_tiles so that
it can be ordered, all_tiles is changed from a set to a vector.

R=reveman@chromium.org
BUG=none

Review URL: https://chromiumcodereview.appspot.com/14261010