Add more DiskCacheBackendTest.SimpleCache*.
Every backend test that runs on the memory backend was added. As well,
I cleaned up the order of tests and some names in
backend_unittest.cc. As well, BackendBasics did not pass due to a bug
in the way we handled the index, now fixed.

R=pasko,felipeg,pliard
BUG=None

Review URL: https://chromiumcodereview.appspot.com/14242015