Better overflow in Type Info section of about:sync
When a per-datatype unrecoverable error occurs, the text associated with
it is displayed in the "Group Type" column of about:sync.  This
description can contain long strings of text without whitespace, and can
cause the table to overflow its div.

This change adds an overflow-x:auto property to the containing div so it
will display a scrollbar if the table outgrows the space allocated to
it.

BUG=236694

Review URL: https://chromiumcodereview.appspot.com/14721003