Create a frame tree in WebContents on the browser side.
The first step to supporting out-of-process iframes is to replicate the frame tree for each page to the browser process. This CL adds a FrameTreeNode class and forms a tree of such nodes in WebContents. Follow up changes will utilize this class to hold frame-specific functionality.

BUG=235879

Review URL: https://chromiumcodereview.appspot.com/14374002