bookmarks: Add a separator to the combobox of the Bookmark Bubble.
With this patch we now show a separator between the most recently used folder
and the "Choose Another Folder..." item in the combobox present in the Bookmark
Bubble when the user press either the star button in the location bar or
Ctrl+D.

BUG=231141
TEST=
1- Press the star button in the location bar to bring the Bookmark Bubble (or
Ctrl+D).
2- Open the combobox.
3- Observe that there is a separator line before the "Choose Another Folder..."
item.
R=erg@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/13881012