Teach the ProxyLauncher to kill() child processes in addition to the browser.
The startup performance_ui_tests have been hanging on the perf bots since they
were upgraded to Precise. The hang is due to the performance_ui_tests binary not
killing all of its child chrome processes. When that happens, the buildbot step
hangs until timeout waiting for the process group to end.

I found that simply sending the children a SIGTERM like we do to the browser
process causes them to exit cleanly. So I'm not sure what the root cause is to
child processes occassionally being left running after the browser exits, but
it seems reasonable to me for the ProxyLauncher to be robust against that case
rather than letting the buildbots hang.

BUG=235893
TEST=Startup performance_ui_tests

Review URL: https://chromiumcodereview.appspot.com/14559003