Files.app: Refresh the tree even when the top-level directory is changed.
When the directory structure is changed, the tree should reflect it immediately.

Previously, when a non-top level directory is changed, the list is changed as well. But when a top-directory is changed, the list is not changed. This is because a bug of the recursive strategy in DirectoryTreeUtil.updateChangedDirectoryItem(). This patch fixes it.

BUG=235107
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/14677005