[Android WebView] Unflake ContentViewZoomTest#testZoomControls
The issue was due to a possibility of a delay in page scale factor
initialization, which lead to a situation when we were trying
to display zoom controls while zooming wasn't yet allowed.

Resolved by synchronizing on the page scale update after page load.

Also, just in case, made zoom controls visibility to be retrieved
on the UI thread in order to avoid races.

BUG=153522
R=benm@chromium.org
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/14850005