Fix: 'You have no extensions' message overlapping installed extensions.
If the ID specified as a query parameter doesn't match any extension in the list, there is a JS error when calculating the offset of the highlight, which leads to the message not being hidden. This fixes the JS error so the execution flow can continue normally.

BUG=224191

Review URL: https://chromiumcodereview.appspot.com/14852008