Fix background tab scrolling issue after r197091
When a tab is opened in background, it's VSyncProvider will be set after
setVSyncNotificationEnabled(true). After VSyncProvider is set, if a
setVSyncNotificationEnabled(true) and a setVSyncNotificationEnabled(false)
are called, the vsync listener will be incorrectly unregistered.

Count the setVSyncNotificationEnabled() calls before VSyncProvider is set.

BTW, also register the listener when setting VSyncProvider if
mVSyncSubscriberCount > 0.

BUG=237011
R=bulach@chromium.org, skyostil@chromium.org, tedchoc@chromium.org

Review URL: https://codereview.chromium.org/14569009