There is a range failure when an old file differs from the file about to be generated on a line containing just a '*'.
This CL fixes that issue by having a length check on the appropriate lists.  It also cleans up the naming of the various variables to make it more obvious which ones refer to the old text file and which ones refer to lines to be written.

NOTRY=true
R=binji@chromium.org, sehr@chromium.org
+sehr for Owners

Review URL: https://codereview.chromium.org/14686003