Handle decoding of vorbis files better on Android
BUG=227144

Send the data size to MediaCodec so we can get the duration length for ogg/vorbis files correctly instead of returning a huge number.  With this information, we no longer need the is_vorbis flag since the duration is now correct, so we remove that code as well.

Review URL: https://chromiumcodereview.appspot.com/14522002