Layout toasts at bottom-left for ChromeOS/RTL UI
ChromeOS will layout the whole desktop in mirrored manner if the
UI language is RTL, so it's more consistent to place toasts at
bottom-left instead of bottom-right.

Windows will keep the bottom-right position, since it doesn't change
the desktop layout itself. See the screenshot of the bug for the
details.

BUG=236988

Review URL: https://chromiumcodereview.appspot.com/14837002