Sanity check to make sure ChannelEstablished handling is valid
If gpu process is compomised, we might receive a ChannelEstablished message but no requests can be found in queue.

Currently we assume there must be a request in queue, leading to potential corrupted function pointer.

The fix in this CL is to check the queue and log a warning and return early if the queue is empty.

I am wondering if a more robust solution is to have request ID and then the ChannelEstablished message should contain the request ID so we could double check. But that's not in the scope of this CL.

BUG=236631
TEST=attached repro.patch in crbug
R=kbr

Review URL: https://chromiumcodereview.appspot.com/14856009