[net] DirectoryListerTest: Operate within a temporary directory structure.
Previously the DirectoryListerTest would work within an existing build directory. This made it difficult to work in isolation, as well as led to flake and slow execution times for the recursive test.

Now, the DirectoryListerTest creates a root ScopedTempDir directory, and builds a directory tree of depth 3 and branching factor of 4. This still tests the recursive descent issues, but runs much faster and requires less test setup.

BUG=155835

Review URL: https://chromiumcodereview.appspot.com/14644009