Instant Extended: Don't resize location entry on suggest
When suggest text is shown the location entry is shrunk.
This means that clicking on the suggest text or the white
space next to it doesn't focus the location entry.

This CL fixes that by not shrinking the location entry when
suggest text is shown. The suggest label ignores mouse
events so clicks correctly go to the location entry.

Note, the location entry is still resized when the suggest
text needs to be on the left (for example, left to right
text in RTL mode).

BUG=224804
TEST=
  Run chrome with instant extended enabled.
  Type "hello w" in the omnibox. Wait for "orld" to be displayed.
  Click on the search results page.
  Click on "orld" and verify that focus switches to the location
  entry and caret is positioned at end of "hello w".

Review URL: https://chromiumcodereview.appspot.com/14145007