Add --enable-bitstream-converter to media/tools/demuxer_bench.
Not only am I interested in measuring the overhead of bitstream conversion, it also makes for a very easy way to reproduce the CHECK() for bug 234449.

Finally, converted a few scoped_ptr<T> to T now that less of our code relies on RefCountedThreadSafe.

BUG=234449
R=dalecurtis@chromium.org

Review URL: https://codereview.chromium.org/14736006