Fixed line height of list notifications on Windows.
Previously this height was 16 pixels. It should have been and is now 18
pixels, the same as the height of message lines.

The baseline-to-baseline distance between the last title line and the
first message line and between the last title line and the first item
line on Windows was fixed: It was previously 20 pixels and should have
been and is now 22 pixels.

The baseline-to-baseline distance between title lines on Windows was
also fixed: It was previously 19 pixels and should have been and is now
20 pixels.

Finally, the distance between the baseline of the last message line or
last item line and the bottom of the card on Windows was fixed. It was
previously 11 pixels and should have been and is now 12 pixels.

BUG=235172
R=dewittj@chromium.org

Review URL: https://codereview.chromium.org/14867002