Pass the hashed username to the NetworkConfigurationUpdater.
This change also simplifies the parameters that are passed to the
NetworkConfigurationUpdater when user policy is ready. Eventually a
NetworkConfigurationUpdater will be created for each Profile, and these
parameters will be passed directly in the constructor.

The hashed username will be used in a subsequent CL (https://codereview.chromium.org/13957012).

BUG=157696

Review URL: https://chromiumcodereview.appspot.com/14771010