Build libpeerconnection as a loadable module for win, mac and linux by default. Also remove the allocator shim variable since it's no longer needed.
This requires updating .isolate files for some try bots.
In practice, I've only seen failures in interactive_ui_tests because of this change, but I'm being a bit proactive here and making sure that libpeerconnection.so exists with test hosts that have tests that depend on it.

BUG=238207
R=csharp@chromium.org, ronghuawu@chromium.org

Review URL: https://codereview.chromium.org/14907003