accelerated_surface_transformer_win: Improve readback and scaling performance by caching temporary surfaces.
This greatly improves the performance of tab capture on current-model Core i3
and Core i5 systems with Intel HD 3000/4000 graphics, and the default drivers
under Win8. On these target systems as much as 15ms was spent in the
~ScopedComPtr dtor, freeing the temporary surfaces.

Patch is intended to be as low-risk as possible while fixing the performance, since most of this code will go away soon when we do not care about non-Aura Windows.

BUG=235502
TEST=surface_gpu_tests

Review URL: https://chromiumcodereview.appspot.com/14763006