Update try e-mail subject for Libyuv Try Server.
This way the conversations get grouped according to
success/failure in the GMail inbox.

The revision variable is removed, because since Libyuv
has no LKGR functionality it will always be empty.
The timestamp variable seems to be empty on all try
servers, so better remove that too.

TEST=none
BUG=none

Review URL: https://chromiumcodereview.appspot.com/14846023