Simplify parseHtmlSubset to use an inert document.
This is a followup on r195976. It reverts most of the changes from that and uses
an inert document instead of removing the attributes like it did before.

A document without a view is specced to be inert. The way to create an inert
document is to use document.implementation.createHTMLDocument('').

BUG=234689

Review URL: https://chromiumcodereview.appspot.com/14741008