[NaCl SDK] nacl_io: Fix dlopen'ing a file not in the .nmf
The dlopen example demonstrates opening a .so file that is not located in the
manifest; for example, from a HTTP mount.

The nacl_io code was broken as a result of a recent change to "fast path" nexe
loading/validation in certain cases. This change attempts to mmap the nexe
using PROT_EXEC, which we would ignore and happily write to a bad address.

Now, we always fail PROT_EXEC requests, which should fallback gracefully to the
old way.

BUG=none
R=sbc@chromium.org
R=ncbray@chromium.org

Review URL: https://codereview.chromium.org/14642024