updating to m29. Misc facts about 29:
29 is the atomic number of copper.
"$29.00" is a song on the album Blue Valentine by Tom Waits.
29 is a prime number.
29 is a Lucas Number.
The Human skull is made-up of 29 bones.
The alphabet of the Phoenicians had 29 cuneiform signs.
The Turkish and Norwegian alphabets contain 29 letters.
Saturn takes 29 years to orbit our Sun.
In the United States of America, Interstate 29 is a freeway 
   which runs from Missouri to North Dakota. 
Also in the USA there is a place in California called 29 Palms.
The probability of being born on 29th February is 1 in 1,461.
A polygon made up of 29 sides is called an icosikaienneagon, enneacosagon or a nonacosagon.

R=laforge@chromium.org

Review URL: https://codereview.chromium.org/15017011