Add cpython 2.7.4 for Windows to third_party.
This is an intentionally very stripped down version of cpython and not 
activestate python. It is 15.9mb versus 87.3mb for python_26 (!) 
(Both including .svn directories) 

Any missing library (like win32api) can be added afterward if it is 
determined that it is truly necessary. 

I intent to do a rolled upgraded, e.g. only new installations get the 
new version and current installation stay on the old version, 
so any issue coming up with this version can be fixed in a controlled manner. 

R=iannucci@chromium.org,szager@chromium.org 
BUG=

Review URL: https://codereview.chromium.org/14962003