[DevTools] Do not unnecessary update devtools split position.
We update split position to the stored value, which may be wrong if nothing
changed with devtools window (split position is stored on dockSide/contents
changes).

This caused a bug that split position was reverted on inspect action.

Previously, we updated split position even if nothing else changed with
devtools window. The single usage of this was implementation of
|changeAttachedWindowHeight|, which only needed to update the height (split).
As it's not used anymore, we can safely remove split positon updating.

Another problem specific to gtk is unnecessary updating split position
on BrowserWindow::Show(), which is fixed now.

BUG=225705
R=pfeldman@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/13912015