Use GetCursorPos to return the current cursor position in windows ASH instead of relying on the last mouse coordinates.
This is needed because some of the real estate in Windows 8 ASH belongs to Windows, i.e. we don't receive mouse
events for portions of the top/top left/top right/bottom parts of the ASH screen.

This fixes the issue with not being able to exit from the fullscreen mode in ASH.

BUG=179988
R=sky@chromium.org
TEST=Manual for Windows ASH at this point. There is work underway to run browser tests for ASH which has fullscreen
     tests.

Review URL: https://codereview.chromium.org/15047004