bookmarks: Make BookmarkModel::IsLoaded() not virtual.
BookmarkModel is not intended to be mocked out nor inherited from. Making
IsLoaded() a virtual function allow clients to inherit from BookmarkModel and
override its behavior, but that is not what we want and thus we make it
non-virtual.

Since r184651, nobody inherits from BookmarkModel nor override IsLoaded().
Also while at it, rename it to just loaded() and inline it in the header file,
since it's a only a trivial getter.

BUG=None
R=sky@chromium.org

Review URL: https://codereview.chromium.org/15012013