Fixes a crash bug happening with --disable-rich-notifications
The code was okay, but in M28, MarkSinglePopupAsShown will invoke
OnNotificationUpdated() which invokes UpdateBubbleView(), and
UpdateBubbleView() calls DeleteTimer() for unused ones. Also
UpdateBubbleView() may delete this if there are no popup notifications,
it is not safe to call DeleteTimer() after MarkSinglePopupAsShown.

This is new to M28, not in M27. Purely a fix for those who
wants to disable rich notifications explicitly.

BUG=238710

Review URL: https://chromiumcodereview.appspot.com/14786016