Only exit GPU process on context lost if explicitly requested via command line.
Previously the GPU process would exit on context lost if either there was no --gpu-driver-workarounds command line switch or the switch was present and the workaround was specifically requested. Now exit on context lost does not happen unless the switch is present and the workaround is requested.

This is a more appropriate default for unit tests since exiting a unit test process prevents subsequent tests from running and does not cause the test that caused the exit to report a failure.

Review URL: https://chromiumcodereview.appspot.com/14663008