Fix the problem of having a dangling layer_ pointer in RenderWidgetFullscreenPepper in some cases.
When PluginInstance calls Destroy() on fullscreen_container_ and resets the pointer, the corresponding RenderWidgetFullscreenPepper instance will post itself a Close message and live until that is processed.  Before it actually goes away, if PluginInstance::UpdateLayer() is called, we may destroy web_layer_ without notifying the soon-to-die RenderWidgetFullscreenPepper instance to reset its layer_ pointer (because we have reset the pointer to it). This will result in a crash if a Resize message is processed before Close.

BUG=236473
TEST=None

Review URL: https://chromiumcodereview.appspot.com/14960005