Make app launcher signin page closer to the mocks.
This rearranges the text a bit, uses different fonts.

The background is also drawn differently. Normally
the background includes a white area and a separator
for the search box. This is now only drawn when the
normal app list view is being shown. If the signin is
shown, the standard background color is used to fill
the entire background.

BUG=159733

Review URL: https://chromiumcodereview.appspot.com/14769015