cc: Run all LayerTreeHost tests with a delegating renderer.
Unless they choose to opt out.

The VideoResourceUpdater was deleting resources when the ResourceProvider
shut down, which was not ideal. Instead, it deletes them when
it shuts down itself (before the ResourceProvider). This works
out fine, since anyone else using it will be using it through a
mailbox.

DelegatingRenderer needs to use and export the resources in the
quads it is given inside of DrawFrame() as they may be deleted
afterward, instead of waiting until SwapBuffers().

Depends on: https://codereview.chromium.org/14960006/

TEST=cc_unittests
R=jamesr, piman
BUG=123444

Review URL: https://chromiumcodereview.appspot.com/14925009