[Android] Always set "mobile user agent" switch to the right vlaue when launching the browser process.
This fixes the user agent for the browser process when the browser process has been started by an activity other than Main, like sync, bookmark provider, etc. If
chrome is started as a background process, the switch is never added. Since the
user agent is computed early on during start-up and not recomputed (from users
of content::GetUserAgent) this can lead to incorrect UA being reported and
broken results.

We probably shouldn't do it this way long-term but this is a minimal fix for m27.

NOTRY=true
BUG=237471

Review URL: https://chromiumcodereview.appspot.com/14682021