Don't send a resize ack when resizing to 0x0
I accidentially broke this in r198991.

I wonder whether it was better to just have the browser tell the
renderer whether it's waiting for an ACK or not. Currently, the browser
always requests an ACK but only expects one if the size changed to a
non-empty rectangle.

BUG=239439
R=piman@chromium.org

Review URL: https://codereview.chromium.org/14620015