Commit InvalidationService implementations
This commit introduces a new ProfileKeyedService, the
InvalidationService.  Clients may register with this service to receive
notifications from a server informing them when they need to fetch
updates to their cached state.  (In other words, they will receive
notifications when their cache has become invalid.)

There are three different implementations of this service, all of which
inherit from InvalidationFrontend.  The current provider of
invalidations, ProfileSyncService, also inherits from this interface.

The TiclInvalidationService receives notifications through the C++
version of the googlecacheinvalidation (aka. TICL) library.  It is used
on Chromium for Desktop operating systems.

The AndroidInvalidationService receives notifications by listening to
the in-process NotificationService
(content/public/browser/notification_service.h).  The code that actually
sends these notifications is a small bridge from Java-land, which is not
affected by this CL.

The P2PInvalidationService uses XMPP peer-to-peer connections to send
and receive invalidations.  This has not been used in Chromium for a
very long time.   We keep it around because some of our integration
tests rely on it.

This CL also includes an InvalidationServiceFactory to help manage the
selection and construction of these InvalidationServices.

The tests for these new services are based on a modified version of
sync/notifier/invalidator_test_template.h.  The P2PInvalidationService
is not tested at all because it is itself a test-only utility.

In addition to the new code, this CL moves around some old code to help
populate a new directory, chrome/browser/invalidation.  We're trying to
separate invalidations from sync, so the
chrome/browser/sync/invalidations directory was no longer needed.

Finally, note that this commit does not change Chromium in any way.  It
adds new classes and test for them, but the rest of the code has not
been modified to make use of them.  Those changes will happen in a
future CL.

BUG=124137

Review URL: https://chromiumcodereview.appspot.com/13991017