Add GetFullRequestHeaders, from URLRequestJob to HttpNetworkTransaction.
Dev Tools displays the raw request and response headers.  Right now, it gets
them by snarfing them out of the NetLog, which is bad, as the NetLog is for
huamn consumption only.  I'm trying to refactor the DevToolsNetLogObserver
away; providing an alternate, supported path to get the request headers is the
first step.

BUG=196304
TEST=added to URLRequest, HttpTransaction, and HttpNetworkTransaction unittests

Review URL: https://chromiumcodereview.appspot.com/12621011