Archive unstripped nacl irt for official builds.
Adding archival of unstripped nacl irt binaries.
(NOTE: We currently do only strip-debug, so symbol info is actually currently
available. Separately we could trim the IRT slightly be dropping this once we
are confident in the archiving.)

This change dependes on this one landing in nacl and being DEPSed in:
https://codereview.chromium.org/14713006/

BUG=None
TEST=None
R=mark@chromium.org,sbc@chromium.org,igorc@chromium.org

Review URL: https://chromiumcodereview.appspot.com/14705005