CrOS Files.app: Hide the middle bar by default on initialize.
This patch hides the middle panel and the splitter at the initialization time. This prevets the middle bar glitch (see crbug.com/239155) when the default directory is on non-drive volume.

Even if the default directory is Drive, the new glitch is not happened by this patch. The middle bar will appear at the same timing as the list updated.

BUG=239155
TEST=manual

NOTRY=True
# NOTRYing for chromeos_linux already passes

Review URL: https://chromiumcodereview.appspot.com/14870005