Network error pages:  Minor UI fixups.
Adjust button sizes/colors, move error code below the fold,
and remove redundant text.  Only "SUMMARIES" are shown in
the full error page, only "DETAILS" in iframes.

Also gets rid of code to use jscontent as the class name,
as it was interfering with some of the new css and needed
to go, anyways.

BUG=174194,239502

Review URL: https://chromiumcodereview.appspot.com/15007005