[Android] Only reinstall test apk if needed
When running tests, we should only be reinstalling the test apk if it
is different than the one currently installed. So, before reinstalling,
query the package manager for the path to the installed apk, and then
only reinstall if the md5sum of the installed apk differs from the one
we are about to install.

NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/14601004