Fix determination of web contents modal dialog parent when WebContents is not visible
Allow WebContentsModalDialogHost to provide the parent window for the
web contents modal dialog. This is where the responsibility for knowing
which window to use most naturally lies, and it avoids the problem of
being unable to determine the parent when the WebContents is not visible
at dialog creation.

BUG=157161, 238392
R=ben@chromium.org, erg@chromium.org, sail@chromium.org

Review URL: https://codereview.chromium.org/15012017