Fixed loading images in Files.app's slide view due to change in Blink.
Recently in Blink onerror events became asynchronous instead of synchronous. As a result loading images in Files.app's slide view stopped working. This was because we were setting the src to an empty string to force reloading of the image and since the onerror handler was empty, there was no problem. After onerror got asynchronous, it became called later, after we assigned an onerror handler for the real image. This caused calling an onerror event for setting an src attribute to en empty string, which was handled as an error in loading the real image.

This patch fixes this issue by using an 1x1 image instead an empty src attribute and waiting until it is loaded before processing the real image.

TEST=Check if Files.app's slide view is working.
BUG=240751
R=yoshiki@chromium.org

Review URL: https://codereview.chromium.org/14589019