Don't expect ack for ViewMsg_OnResize if backing size is empty
The renderer won't send ack for ViewMsg_OnResize if backing size is empty.
This happens on Android in some cases, e.g. when a tab is created in
background because the backing is only created when the tab is shown.

Previously on Android we let RWHV return empty view bounds when
backing is empty but it prevents the renderer from getting the correct
view bounds. Now remove the above logic from RWHV_android and change
RWHI not to expect ack when backing size is empty.

Also changed RenderWidget not to send ack on the next paint in the cases
that the host is not expecting the ack. This doesn't change the actual
behavior because RenderWidget won't send UpdateRect corresponding to the
Resize request in the cases, but just ensures the correctness.

This is a part of the fix to bug 168568. Also need to change the Java code
to set the correct view size as early as possible.

BUG=168568
TEST=RenderWidgetHostTest.Resize (content_unittests), RenderWidgetTest.OnResize (content_browsertests)

Review URL: https://chromiumcodereview.appspot.com/14779010