Fix crash when dropping crx to chrome:extensions from Google Drive on Chrome OS.
WebDropData has two fields .path and .display_name.
The .path field points to the real path to the file, but it can be a random
hash value or something like that when the file is mounted through HTML5
file system. In order to check the file name extension, .display_name
containing a user-friendly name representation has to be used.

BUG=238495
TEST=Manually tested drag&drop on CrOS and non-CrOS build.

Review URL: https://chromiumcodereview.appspot.com/15089003