Don't leak HostGlobals in PpapiUnittests
In PpapiUnittest we construct a PluginModule which creates a leaky HostGlobals.
When running unittests, if a test constructs its own PpapiGlobals after a
PpapiUnittest has run, it will hit a DCHECK which asserts that no PpapiGlobals
already exists.

This CL tears down the HostGlobals after the PpapiUnittest has run so that there
is no test interference. We should have a better way to avoid leaking
HostGlobals in production code if possible so I added a TODO.

R=yzshen@chromium.org

Review URL: https://codereview.chromium.org/14642027