Window placement for the OSX App Launcher.
This places the window inside the work area. It is snapped to the dock
if the dock is on the same screen as the mouse cursor. Otherwise it is
placed at the bottom left of the screen.

BUG=138633
TEST=Launch the OSX App Launcher with the dock in various positions, and
the cursor in various places on the screen. The launcher should appear
near the dock, or in the bottom left of a screen without a dock.

Review URL: https://chromiumcodereview.appspot.com/14874025