Refactor OSExchangeData::CustomFormat to use Clipboard::FormatType.
This was previously a typedef for a platform-specific type, and it
could make writing cross-platform code harder; for example, see
WebContentsViewAura, which needed to to use #if statements so it'd
compile on all supported platforms. In addition, there was a strange
overlap between the functionality of CustomFormat and FormatType.

Consolidating them helps clean up several areas:
- Format definitions for Windows were previously duplicated between
  ClipboardUtilWin and Clipboard. Now they only need to be defined in
  Clipboard.
- OSExchangeDataProviderWin had some situations where it extracted
  cfFormat from a FORMATETC and then reconstituted a FORMATETC. It has
  been updated to just use a FORMATETC (which can be extracted from a
  FormatType).
- OSExchangeData::RegisterCustomFormat has been removed in favor of just
  using Clipboard::GetFormatType directly.
- IDataObject takes a FORMATETC*, which is non-const. For better safety,
  the code has been updated so that IDataObject can never modify the
  "canonical" FORMATETC.

R=sky@chromium.org

Review URL: https://codereview.chromium.org/14949014