Handle exception from getDaemonVersion on NPAPI plugin.
This exception-handling was mistakenly removed in r199035.
This CL restores it, which is sufficient to allow the webapp
to start up properly when there is no NPAPI support.

Currently, every method of HostDispatcher can potentially
throw - these will all be fixed as part of a general
error-handling cleanup.

BUG=240000
TEST=remoting webapp loads on ChromeOS
R=jamiewalch@chromium.org

Review URL: https://codereview.chromium.org/14766013