Move webkit/compositor_bindings into webkit/renderer
The compositor bindings for WebKit should only be used in a render process or in
webkit_support for DumpRenderTree or webkit_unit_tests. It's never used in a
browser process. This moves the bindings code under webkit/renderer/ so we can
enforce this via DEPS rules.

The path is unfortunately a bit awkward. Once we drop DumpRenderTreeSupport and
figure out a way to deal with webkit_unit_tests, we can move the bindings into
content/renderer/gpu/ since they'll only be used by the content embedding.

BUG=239107
TBR=piman@chromium.org

Review URL: https://codereview.chromium.org/14651027