cc: Try eliminate flake on a cc unittest.
LayerTreeHostTestAtomicCommitWithPartialUpdate flaked on the
waterfall today. It looks like it drew an extra frame somehow
with no damage, but I can't see how it would. So let's try
make the test robust to that anyway, and see if that fixes it
or gives us more data.

R=enne
BUG=240757

Review URL: https://chromiumcodereview.appspot.com/15123006