Add support for a cookie retention priority field trial.
For the experiment group, a new algorithm will be used to identify the cookies to be evicted when a given domain exceeds its limits. The new algorithm will only differ from the old algorithm when a given domain uses the 'priority' attribute in its cookies.

BUG=241178

Review URL: https://chromiumcodereview.appspot.com/14772046