Remove sync-related login code on OS_CHROMEOS platforms.
ChromeOS doesn't need the sync login code - in fact, it's actively bad to have
this code in there because it keeps getting triggered through various bugs. As
an interim measure, this CL #if-outs the code on chromeos platforms. Once we
have removed the non-web-based signin code, we can look at the remaining code
and decide whether it is worth refactoring this into platform-specific classes.

BUG=236005
R=arv@chromium.org, rogerta@chromium.org

Review URL: https://codereview.chromium.org/14696017