Fix lookup of unqualified name into template dependent base class
C++ requires that you qualify names inherited from a dependent base.
MSVC allows these name lookups to succeed, but clang does not.  The
usual way to fix the lookup is to prefix the member names with this->,
but this is not common in Chrome, so I added some using decls to pull in
the member names.

Clang actually fails to diagnose this properly with -fms-compatibility,
and I ended up with a link error.  The upstream bug is
http://llvm.org/PR16014.

R=cpu@chromium.org
BUG=82385

Review URL: https://codereview.chromium.org/15120008