Add ContentViewCore.setInitialViewportSize()
When a page is opened in a background tab, it will be loaded but the
viewport size is only initialized when the tab becomes visible. This causes
the page to see 0x0 viewport size and may be broken (e.g. div by zero).

Add ContentViewCore.setInitialViewportSize() so that the app can set reasonable
initial viewport size parameters before the ContentViewCore get the real ones
from onSizeChanged().

BUG=168568
R=joth@chromium.org

Review URL: https://codereview.chromium.org/15067006