Parse WebM track name and language
The Track header of a WebM file contains (among other items)
a track name and track language. We do not currently use
this information. However, it is needed to construct WebVTT
inband text tracks. The WebM tracks parser has been modified
to provide this information to its clients.

(An earlier version of this commit failed to compile on some
systems. The problem was that some integer constants needed
to be qualified with a suffix to specify an extra precision
integer type.)

BUG=230708

Review URL: https://chromiumcodereview.appspot.com/15205002