chrome://nacl asserts
A check was added in r180577 to check that the pnacl directory exists. The check runs in the UI thread and calls file_util::PathExists().
The UI thread does not have permission to make this call, and it asserts in ThreadRestrictions::AssertIOAllowed
Create a NaClDOMHandlerProxy class that performs this check in the FILE thread.

BUG=240013

Review URL: https://chromiumcodereview.appspot.com/14860020