Implement server side copy of regular files.
Drive API v2 supports to copy regular files on server side, which is not on
GData WAPI. This CL changes to use the feature only when Drive API v2 is
enabled. By this change, we don't need to download and then upload the file
on Drive API v2.

BUG=138273
TEST=Ran unit_tests and tested manually.

Review URL: https://chromiumcodereview.appspot.com/15085022