Implement serving real data from the FakeDriveService.
To implement more advanced end-to-end tests we need real files, such as images or even movies. The previous FakeDriveService implementation was returning a generated string of desired length consisting of only one character. This patch allows to pass the file data we want to be returned.

TEST=unit_tests: FakeDriveServiceTest*
BUG=241022

Review URL: https://chromiumcodereview.appspot.com/14676019