[sync] Separate sign in from sync on Desktop Chrome
As of today, when a user disables sync on desktop chrome, we
automatically sign out too. With recent changes that have gone in under
the covers, this is no longer necessary. It is possible for the user to
choose not to sync any datatypes, and still be signed in to chrome.

This patch does multiple things:
- Adds a "Sync nothing" option to the drop-down menu in the advanced
  sync setup dialog.
- Detaches sign in from sync by no longer calling SignOut when the user
  disables sync.
- Updates the Sign in / Users section of the chrome settings page and the
  advanced sync settings dialog on Desktop Chrome and Chrome OS.
- Eliminates the re-authentication step when the user enables sync while
  being already signed in.
- Dismisses the advanced sync settings dialog if it is still visible when the
  user gets signed out in the background.
- Adds new strings and modifies existing strings to account for the new
  UI.

BUG=229860, 237051, 230348, 238938, 240436, 128692, 94814
TEST=Exercise the advanced sync settings dialog to enable and disable sync
R=atwilson@chromium.org, estade@chromium.org

Review URL: https://codereview.chromium.org/14691004