InstantExtended: Reload when re-accepting search terms.
Previously when focusing in the omnibox and pressing enter while it
showed search results, we would re-classify the match based on the
query terms. Instead, we should refresh the search page. Not doing so
is a potential security problem and is inconsistent UI behavior, because
something that says "Search" should really do a search.

TEST=Manually and browsertest
BUG=173483,226841

Review URL: https://chromiumcodereview.appspot.com/14752021