Adding new user menu section to the SystemTrayMenu & refactoring of user access
It is still not fully finished - but it is functional and might be useful for what you are doing (and already quite big as it is). Note that the functionality is controlled by the flag.

What is missing?
- There are no unit tests yet
- The visual Drop down box to add a new user is missing.
- The error message box which shows the error if there are already 3 users signed in is missing.
- At the moment I have ~7 lines of test code in there which I will remove before I check in. (I need them since the multi login still does not work for me)

I am not sure if you want to wait until I finish the issue - or if you'd rather want to have this to be able to rather merge early then later. I leave this up to you. If you are for the latter one, you can have at least a look at my changes and if they conflict with yours.

BUG=239201
TEST=visual

Review URL: https://chromiumcodereview.appspot.com/14756019