[sync] Remove the abomination that is ProfileSyncService::ResetForTest
ProfileSyncService::ResetForTest was originally added as a means to
recreate a PSS object in place, so that a sync integration test could
simulate browser restart without actually destroying and recreating a
browser instance.

Unfortunately, this was always prone to breakage, and involved an ugly
placement-new hack. The tests that used this method were disabled as a
result.

This patch removes the method and its associated tests. There are other
means by which an integration test can restart the browser and re-use
the same profile directory, making this redundant.

BUG=none
TEST=sync_integration_tests
R=tim@chromium.org

Review URL: https://codereview.chromium.org/15371002