[NaCl SDK] "make run" should make all, then run webserver.
This was broken recently when I tried to use MAKECMDGOALS to generate the same
target for "make all" and "make clean". Because "make run" depends on the all
target, it was recursively running "make run" on all subdirectories.

This CL also fixes the path to httpd.py.

BUG=241610
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/15331004