Let Android port access properly sized images
Depends on assets in http://crrev.com/15136005, which had to be separated because of the trybots.

With the way things are currently set up, Chrome on Android can't access properly
sized image resources:
* The APK only packages the mostly unused 100% resources (none of
  the larger 200% assets)
* The bitmaps don't scale properly when they're converted to Java bitmaps because
  the Java code knows nothing about the resource's expected dimensions on different
  devices
* The Chromium resource loading system doesn't allow us to take advantage of
  Android's resource loading schemes (drawable directories for different screen
  sizes, RTL support, etc.).

Instead of #ifdefing out practically everything in the theme_resources.grd file,
this CL introduces a map between the chromium resource IDs and Android drawables,
allowing us to load up the correctly sized Android assets.

The template system is bootstrapped with the InfoBar and WebsiteSettingsUI icon
assets.

TBR=jochen
BUG=237034,238668

Review URL: https://chromiumcodereview.appspot.com/14772036