Fix check for updating WebDragDestGtk when creating a view for a RVH.
CreateViewForWidget can be called for navigations that don't end up
swapping in the render view. An example of this would be dragging and
dropping a file on Chrome that it can't display, such as an executable.
This is problematic because it creates a WebDragDestGtk for a view that
is never actually displayed, thus breaking all drag and drop operations
until that tab is navigated. To fix this, it's not sufficient to just
check that the RVH isn't swapped out. Instead, update WebDragDestGtk
only if the RVH is already the RVH for the web contents; otherwise,
allow RenderViewSwappedIn to update it.

BUG=241564

Review URL: https://chromiumcodereview.appspot.com/14918012