Fix bug in GetAbsoluteWebStringFromUTF8Path() introduced by refactoring MakeAbsoluteFilePath()
r193855 replaced 'bool AbsolutePath(base::FilePath*)' with 'FilePath
MakeAbsoluteFilePath(const FilePath&)'.

Both functions fail if an error occurs (for example the path does not exist).
In the case of AbsolutePath(), the input FilePath is unchanged and false is
returned. In the case of MakeAbsoluteFilePath(), an empty FilePath is
returned.

This change in behavior caused 10 LayoutTests to fail, so this patch restores
the previous behavior.

BUG=230771
R=brettw@chromium.org, tkent@chromium.org

Review URL: https://codereview.chromium.org/14634008