Post integrity_check data for corrupt thumbnail databases.
Analyzing the results from DatabaseErrorCallback() show a lot of
SQLITE_CORRUPT opening the database and setting up the meta table.
Add additional diagnostic information for these cases to direct future
development, and also dial down the amount of data uploaded to allow
finding other cases more easily.

For SQLITE_CORRUPT and SQLITE_READONLY, only give the chance to report
on first encounter.  Cross-platform results are inconsistent.  On OSX,
almost all of the SQLITE_CORRUPT cases are in initialization, whereas
on Windows there are more in code which updates the database.  For
these errors, it is reasonable to expect any future call to return the
error again, so it is pointless to let it get multiple attempts.

BUG=240396

Review URL: https://chromiumcodereview.appspot.com/15327004