Rename test callbacks for CookieStore to represent the argument type.
These callbacks are not tightly connected to each method but are made to save
a result value of the specified type. Rename them to represent the type.

Actually, SetCookieCallback is used for testing Delete method which passes
a bool result to the done callback.

Review URL: https://chromiumcodereview.appspot.com/15380003