[Telemetry] Move discover.py to core/ and add unit test
discover.py is used in several places in telemetry unrelated to the unit tests, move it to core/.

This CL also adds a telemetry.test.test_data package.  The motivation for this is that the profile creators code also wants to test the import of classes so it's useful to have a package that we can pass to discover.

BUG=None
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/15484009