Translate device-local account IDs to user IDs
This CL decouples the namespaces used for device-local accounts in policy
and user management. In addition to the account ID assigned by the server,
each device-local account is given a user ID that always follows a
canonical format and allows the account type to be identified just by
looking at the user ID, regardless of the format chosen for the account ID
by the server.

BUG=240276
TEST=Manual in VM

TBR=jhawkins (chrome/browser/ui/webui/policy_ui.cc)

Review URL: https://chromiumcodereview.appspot.com/14927015