Improves capture delay estimates on Mac.
This CL ensures that we take the browser-side FIFO into account when we derive the capture delay. Before, a typical delay estimate given to the renderer was about 3ms, now with the new patch, this value is 10-12 ms which is more inline with the 10ms buffer size used by e.g. WebRTC.

BUG=none
TEST=Internal logging using different WebRTC clients on Mac.
R=xians@chromium.org

Review URL: https://codereview.chromium.org/15724002