Avoid an infinite loop in drive::ResourceMetadata.
It may happen that the while loop is infinite if the base name is empty or a single or two dots. This may happend where entries are manually added or if the feed is malformed. This patch adds a DCHECK which checks if the alternative filename is different than the previous one.

TEST=Tested manually.
BUG=242424

Review URL: https://chromiumcodereview.appspot.com/15557002